Additional load: persist history via storage and split into its own module
- Move all named additional-load forecast handling out of fetch.py, plan.py
  and output.py into a new additional_load.py mixin (AdditionalLoad)
- Persist completed-load history via the storage component instead of a HA
  sensor attribute; the sensor now publishes only the count and a recent
  preview to avoid recorder bloat
- Add configurable house_load_additional_flexible_max_hours to bound the
  flexible candidate search (defaults to the forecast horizon)
- De-duplicate the additional_load_forecast_record call sites and document
  that load fields may reference HA entities
