+++ apps/predbat/plan.py
@@ -189,193 +189,6 @@ class Plan:
     one C++ kernel call, which spreads it across its own threads.
     """
 
-    def additional_load_candidate_profile(self, forecast, start_minutes):
-        """
-        Build absolute-minute adjustment and target metadata for one flexible load candidate.
-        """
-        plan_interval = forecast.get("plan_interval_minutes", self.plan_interval_minutes)
-        duration_minutes = int(forecast.get("duration", 0.0) * 60)
-        end_minutes = start_minutes + duration_minutes
-        periods = forecast.get("_periods", 0)
-        weights = forecast.get("_weights", [])
-        weight_total = forecast.get("_weight_total", sum(weights))
-        energy_total = forecast.get("energy", None)
-        slot_energy = forecast.get("slot_energy", 0.0)
-        load_adjust = {}
-        target_times = []
-        total_energy = 0.0
-
-        for period in range(periods):
-            slot_start = start_minutes + period * plan_interval
-            slot_end = min(slot_start + plan_interval, end_minutes)
-            slot_minutes = slot_end - slot_start
-            if slot_end <= self.minutes_now:
-                continue
-            if (slot_start - self.minutes_now) >= self.forecast_minutes:
-                continue
-            energy, adjustment_energy = self.additional_load_slot_energies(energy_total, slot_energy, weights, weight_total, period, slot_minutes, plan_interval)
-            total_energy += energy
-            for minute in range(slot_start, slot_end):
-                load_adjust[minute] = dp4(load_adjust.get(minute, 0.0) + adjustment_energy)
-            target_times.append({"start": (self.midnight_utc + timedelta(minutes=slot_start)).isoformat(), "end": (self.midnight_utc + timedelta(minutes=slot_end)).isoformat(), "energy": energy})
-
-        return load_adjust, target_times, dp4(total_energy)
-
-    def add_additional_load_to_step_data(self, load_minutes_step, load_adjust):
-        """
-        Add absolute-minute additional load adjustment into prediction step data.
-        """
-        modified_load = copy.deepcopy(load_minutes_step)
-        if not load_adjust:
-            return modified_load
-        for minute_relative in range(0, self.forecast_minutes, PREDICT_STEP):
-            minute_absolute = self.minutes_now + minute_relative
-            bucket_energy = 0
-            for offset in range(PREDICT_STEP):
-                bucket_energy += load_adjust.get(minute_absolute + offset, 0.0) / float(self.plan_interval_minutes)
-            if bucket_energy <= 0:
-                continue
-            modified_load[minute_relative] = dp4(modified_load.get(minute_relative, 0.0) + bucket_energy)
-        return modified_load
-
-    def select_flexible_additional_loads(self, load_minutes_step, load_minutes_step10, pv_forecast_minute_step, pv_forecast_minute10_step):
-        """
-        Select flexible additional load start times using full prediction metric impact.
-        """
-        self.house_load_additional_flexible_selection_changed = False
-        flexible_forecasts = {name: forecast for name, forecast in self.house_load_additional_forecasts.items() if forecast.get("enabled") and forecast.get("mode") == "flexible" and not forecast.get("selection_locked", False)}
-        if not flexible_forecasts:
-            return False, load_minutes_step, load_minutes_step10
-
-        selected_flexible = {}
-        working_load_step = load_minutes_step
-        working_load_step10 = load_minutes_step10
-
-        for name, forecast in flexible_forecasts.items():
-            start_minutes = forecast.get("_requested_start_minutes", None)
-            end_minutes = forecast.get("_requested_end_minutes", None)
-            duration_minutes = int(forecast.get("duration", 0.0) * 60)
-            plan_interval = forecast.get("plan_interval_minutes", self.plan_interval_minutes)
-            if start_minutes is None or end_minutes is None or duration_minutes <= 0:
-                continue
-
-            candidate = max(start_minutes, self.minutes_now)
-            candidate = int((candidate + plan_interval - 1) / plan_interval) * plan_interval
-            latest_start = min(end_minutes - duration_minutes, self.minutes_now + self.forecast_minutes - duration_minutes)
-            if latest_start < candidate:
-                continue
-
-            baseline_prediction = Prediction(self, pv_forecast_minute_step, pv_forecast_minute10_step, working_load_step, working_load_step10)
-            baseline_metric = self.score_flexible_additional_load_prediction(baseline_prediction)
-            best_start = None
-            best_metric = None
-            candidate_count = 0
-
-            while candidate <= latest_start:
-                candidate_adjust, _, _ = self.additional_load_candidate_profile(forecast, candidate)
-                candidate_load_step = self.add_additional_load_to_step_data(working_load_step, candidate_adjust)
-                candidate_load_step10 = self.add_additional_load_to_step_data(working_load_step10, candidate_adjust)
-                candidate_prediction = Prediction(self, pv_forecast_minute_step, pv_forecast_minute10_step, candidate_load_step, candidate_load_step10)
-                candidate_metric = self.score_flexible_additional_load_prediction(candidate_prediction)
-                candidate_count += 1
-                if best_metric is None or candidate_metric < best_metric:
-                    best_metric = candidate_metric
-                    best_start = candidate
-                candidate += plan_interval
-
-            if best_start is not None:
-                existing_start = forecast.get("_selected_start_minutes", None)
-                if existing_start is None and forecast.get("suggested_start"):
-                    existing_start = self.additional_load_stamp_to_minutes(forecast.get("suggested_start"))
-                if existing_start != best_start:
-                    self.house_load_additional_flexible_selection_changed = True
-                best_adjust, _, _ = self.additional_load_candidate_profile(forecast, best_start)
-                working_load_step = self.add_additional_load_to_step_data(working_load_step, best_adjust)
-                working_load_step10 = self.add_additional_load_to_step_data(working_load_step10, best_adjust)
-                selected_flexible[name] = {
-                    "_requested_start_minutes": start_minutes,
-                    "_requested_end_minutes": end_minutes,
-                    "_selected_start_minutes": best_start,
-                    "_selection_reason": "prediction_metric",
-                    "_candidate_count": candidate_count,
-                    "_selected_metric": dp2(best_metric) if best_metric is not None else None,
-                    "_baseline_metric": dp2(baseline_metric),
-                    "_expires_minutes": best_start + duration_minutes if forecast.get("auto_expire", False) else None,
-                }
-                if forecast.get("auto_expire", False):
-                    self.house_load_additional_forecast_overrides[name] = {"name": name, **selected_flexible[name]}
-                    self.update_additional_load_api_command_metadata(
-                        name,
-                        {
-                            "_requested_start": self.additional_load_minutes_to_stamp(start_minutes),
-                            "_requested_end": self.additional_load_minutes_to_stamp(end_minutes),
-                            "_selected_start": self.additional_load_minutes_to_stamp(best_start),
-                            "_selected_end": self.additional_load_minutes_to_stamp(best_start + duration_minutes),
-                            "_expires_at": self.additional_load_minutes_to_stamp(best_start + duration_minutes),
-                            "_selection_reason": "prediction_metric",
-                            "_candidate_count": candidate_count,
-                            "_selected_metric": dp2(best_metric) if best_metric is not None else None,
-                            "_baseline_metric": dp2(baseline_metric),
-                        },
-                    )
-                self.log("Flexible additional load {} selected {}-{} using prediction metric {} from {} candidates".format(name, self.time_abs_str(best_start), self.time_abs_str(best_start + duration_minutes), dp2(best_metric), candidate_count))
-
-        if not selected_flexible:
-            return False, load_minutes_step, load_minutes_step10
-
-        self.house_load_additional_forecast_adjust, self.house_load_additional_forecasts = self.fetch_additional_load_forecast(selected_flexible=selected_flexible)
-        self.publish_additional_load_forecasts()
-        return True, working_load_step, working_load_step10
-
-    def score_flexible_additional_load_prediction(self, prediction):
-        """
-        Score a flexible additional load candidate using the same full metric as the optimiser.
-        """
-        (
-            cost10,
-            _import_kwh_battery10,
-            _import_kwh_house10,
-            _export_kwh10,
-            _soc_min10,
-            soc10,
-            _soc_min_minute10,
-            _battery_cycle10,
-            _metric_keep10,
-            final_iboost10,
-            _final_carbon_g10,
-            *_prediction_detail10,
-        ) = prediction.run_prediction(self.charge_limit_best, self.charge_window_best, self.export_window_best, self.export_limits_best, True, self.end_record)
-        (
-            cost,
-            import_kwh_battery,
-            import_kwh_house,
-            export_kwh,
-            _soc_min,
-            soc,
-            _soc_min_minute,
-            battery_cycle,
-            metric_keep,
-            final_iboost,
-            final_carbon_g,
-            *_prediction_detail,
-        ) = prediction.run_prediction(self.charge_limit_best, self.charge_window_best, self.export_window_best, self.export_limits_best, False, self.end_record)
-        metric, _battery_value = self.compute_metric(
-            self.end_record,
-            soc,
-            soc10,
-            cost,
-            cost10,
-            final_iboost,
-            final_iboost10,
-            battery_cycle,
-            metric_keep,
-            final_carbon_g,
-            import_kwh_battery,
-            import_kwh_house,
-            export_kwh,
-        )
-        return metric
-
     def dynamic_load(self):
         """
         Adjust load prediction based on current load
