+++ apps/predbat/predbat.py
@@ -87,6 +87,7 @@ from marginal import Marginal
 from plan import Plan
 from fetch import Fetch
 from output import Output
+from additional_load import AdditionalLoad
 from userinterface import UserInterface
 from compare import Compare
 from plugin_system import PluginSystem
@@ -95,7 +96,7 @@ from ha import run_async
 from control_ledger import ControlLedger
 
 
-class PredBat(hass.Hass, Octopus, Energidataservice, Stromligning, Fetch, Plan, Marginal, Execute, Output, UserInterface, GitHub):
+class PredBat(hass.Hass, Octopus, Energidataservice, Stromligning, Fetch, Plan, Marginal, Execute, Output, AdditionalLoad, UserInterface, GitHub):
     """Main PredBat orchestrator combining all subsystems via multiple inheritance.
 
     Inherits from Hass (HA interface), Octopus (rate loading), Energidataservice, Stromligning,
